requirements.txt
torch==2.4.0
typer==0.15.1
fastapi==0.115.6
uvicorn==0.34.0
pillow==11.1.0
